stock_trader cleaned up. mat does det nows. stock_trader now uses my preferred style for attributes in classes. If the property is a pointer then it should come after basically constants.
The matrix library does determinants, I've also started to get it ready
for the real-deal stuff. I'm going to have a hardcoded 2x2 of the
adjugate and determinant functions since they're easy enough to by hand.
And the hill cipher i'll be working with will only be doing it on a 2x2
matrix.