stock_trader cleaned up. matrix class continues to be worked on. The matrix class will use an in-place inversion method that's similar to LU decomposition so that it can have O(2/3(n**3)) time complexity than the current one that uses the adjugate, and the determinant. That version takes O(2/3(n**3)) for the determinant, then we have to operate on each element of the matrix via the cofactor. So in the end it's really. A much worse time complexity of O(((n-1)**2)*(n**2)) for just the adjugate. Then also have to multiply this adjugate matrix times the determinant. Which will take O(n**2) time.
So our final time for the old method is O((n**2)+(((n-1)**2)*(n**2))).
This compared to LUDcomposition which is just a measily O(2/3*(n**3))
shows a massive speedup. In real times for a 4x4 matrix old method would
take 160 operations whereas this one will just take a measley ~43
operations. Almost 1/4 the amount of operations which should really come
in handy in the end.

This method is still going to be used for modular inverses until I can
find a better way of doing it. But all other inversions will use
LUDecomposition.

Or to put it in concrete numbers. To calculate