hill_cipher.hxx, matrix.hxx, random.h, match.hxx, terminal.h, vectors.hxx all updated.
math.h renamed to math.hxx due to it not working with C at this time.
Hill cipher now has key generation methods and also setters.
random.h now updated to support C code more easily(not complete but it's a skeleton). The seeder also checks to see if it's already been seeded. If it has been and they didn't provide a seed it'll simply exit. Otherwise it'll seed it either with the provided seed or the current microtime.
Due to math.h being renamed a lot of other files were also updated to use this new name.