hill_cipher.hxx updated to include the first part of the data chunker.
It'll read in the input data and then create a matrix the same size as the key and then assign each character of the input to the proper element of the matrix for later. It then returns a std::vector of these Matrix<char>s that are then accessed in order by the encryption method(one of the two the later will literally just multiply the matrices mod alphabet_size). Then return after this it will give you back your data as encrypted matrices. The dechunk_it method will then translate this back into a string for use from sending to the terminal or writing to a file.

The driver program will include code that'll read in files/from stdin line by line.