matrix.hxx working more as intended. matrix_test.cpp more operators.
As stated I've broke out the float type methods for the ones where I did do rounding so that if someone wants to use floats(why they're the same cost as doubles on modern hadware) can do so w/o precision loss. Also the integer types are now being floored as they should be so that the conversions work.