itp232/stock_trader.cpp: Fixed up a lot of UX things. Handled EOF. itp232/stock_trader.h: Made the buy/sell UX better. itp232/terminal.h: Made it so that EOF is handled better now.
Before getting an EOF would cause the thing to end up in an endless loop.
Now it'll actually watch/notice that kind of thing and just exit the
program with a return code of 1.
stocktrader.h
When buying something it tells you how many of an item you have. Also
if you own nothing it makes sure to set that flag.
stock_trader.cpp
If you own nothing when it comes time to sell it'll be sure to tell
you that. We already check that value upon you selling something
but this is here to prevent someone from having yet another
pointless set of options. Also the color scheme is updated to be what
it was always supposed to be all along.